refactor(reports): colocate tests with their feature subdirectories
